requirements.txt
pygame==2.5.2
neat-python==0.92
matplotlib==3.8.4